Updating docs and consolidating args.
