+++ doc/README.open_ce.md
@@ -1,0 +1,5 @@
+# README for `open-ce` tool
+
+* `open-ce build`
+* `open-ce test`
+* `open-ce validate`
