+++ git_tools/create_opence_release.py
@@ -19,9 +19,12 @@ import glob
 import git_utils
 import tag_all_repos
 
+sys.path.append(os.path.join(pathlib.Path(__file__).parent.absolute(), '..', 'open-ce'))
+import utils # pylint: disable=wrong-import-position
+
 def _make_parser():
     ''' Parser input arguments '''
-    parser = git_utils.make_parser([git_utils.Argument.PUBLIC_ACCESS_TOKEN, git_utils.Argument.REPO_DIR,
+    parser = utils.make_parser([git_utils.Argument.PUBLIC_ACCESS_TOKEN, git_utils.Argument.REPO_DIR,
                                     git_utils.Argument.BRANCH],
                                     description = 'A script that can be used to cut an open-ce release.')
