+++ git_tools/git_utils.py
@@ -119,12 +119,3 @@ def ask_for_input(message, acceptable=None):
         print("{} is not a valid selection.".format(user_input))
         user_input = input(display_message)
     return user_input.lower()
-
-def make_parser(arguments, *args, formatter_class=utils.OpenCEFormatter, **kwargs):
-    '''
-    Make a parser from a list of OPEN-CE Arguments.
-    '''
-    parser = argparse.ArgumentParser(*args, formatter_class=formatter_class, **kwargs)
-    for argument in arguments:
-        argument(parser)
-    return parser
