+++ open-ce/utils.py
@@ -8,16 +8,15 @@ disclosure restricted by GSA ADP Schedule Contract with IBM Corp.
 """
 
 import os
-import argparse
 import sys
 import subprocess
 import errno
-from enum import Enum, unique
 from itertools import product
 import re
 import yaml
 import pkg_resources
 from errors import OpenCEError, Error
+import inputs
 
 
 DEFAULT_BUILD_TYPES = "cpu,cuda"
@@ -33,144 +32,13 @@ CONDA_ENV_FILENAME_PREFIX = "opence-conda-env-"
 DEFAULT_OUTPUT_FOLDER = "condabuild"
 DEFAULT_TEST_CONFIG_FILE = "tests/open-ce-tests.yaml"
 
-class OpenCEFormatter(argparse.ArgumentDefaultsHelpFormatter):
-    """
-    Default help text formatter class used within Open-CE.
-    Allows the use of raw text argument descriptions by
-    prepending 'R|' to the description text.
-    """
-    def _split_lines(self, text, width):
-        if text.startswith('R|'):
-            return text[2:].splitlines()
-        return super()._split_lines(text, width)
-
-@unique
-class Argument(Enum):
-    '''Enum for Arguments'''
-    CONDA_BUILD_CONFIG = (lambda parser: parser.add_argument(
-                                        '--conda_build_config',
-                                        type=str,
-                                        default=DEFAULT_CONDA_BUILD_CONFIG,
-                                        help='Location of conda_build_config.yaml file.' ))
-
-    OUTPUT_FOLDER = (lambda parser: parser.add_argument(
-                                        '--output_folder',
-                                        type=str,
-                                        default=DEFAULT_OUTPUT_FOLDER,
-                                        help='Path where built conda packages will be saved.'))
-
-    CHANNELS = (lambda parser: parser.add_argument(
-                                        '--channels',
-                                        dest='channels_list',
-                                        action='append',
-                                        type=str,
-                                        default=list(),
-                                        help='Conda channels to be used.'))
-
-    ENV_FILE = (lambda parser: parser.add_argument(
-                                        'env_config_file',
-                                        nargs='+',
-                                        type=str,
-                                        help="Environment config file. This should be a YAML file "
-                                             "describing the package environment you wish to build. A collection "
-                                             "of files exist under the envs directory."))
-
-    REPOSITORY_FOLDER = (lambda parser: parser.add_argument(
-                                        '--repository_folder',
-                                        type=str,
-                                        default="",
-                                        help="Directory that contains the repositories. If the "
-                                            "repositories don't exist locally, they will be "
-                                            "downloaded from OpenCE's git repository. If no value is provided, "
-                                            "repositories will be downloaded to the current working directory."))
-
-    PYTHON_VERSIONS = (lambda parser: parser.add_argument(
-                                        '--python_versions',
-                                        type=str,
-                                        default=DEFAULT_PYTHON_VERS,
-                                        help='Comma delimited list of python versions to build for '
-                                             ', such as "3.6" or "3.7".'))
-
-    BUILD_TYPES = (lambda parser: parser.add_argument(
-                                        '--build_types',
-                                        type=str,
-                                        default=DEFAULT_BUILD_TYPES,
-                                        help='Comma delimited list of build types, such as "cpu" or "cuda".'))
-
-    MPI_TYPES = (lambda parser: parser.add_argument(
-                                        '--mpi_types',
-                                        type=str,
-                                        default=DEFAULT_MPI_TYPES,
-                                        help='Comma delimited list of mpi types, such as "openmpi" or "system".'))
-
-    CUDA_VERSIONS = (lambda parser: parser.add_argument(
-                                        '--cuda_versions',
-                                        type=str,
-                                        default=DEFAULT_CUDA_VERS,
-                                        #Supress description of cuda_versions flag until more robust testing
-                                        help=argparse.SUPPRESS))
-                                        #help='Comma delimited list of cuda versions to build for '
-                                        #     ', such as "10.2" or "11.0".'))
-
-    DOCKER_BUILD = (lambda parser: parser.add_argument(
-                                        '--docker_build',
-                                        action='store_true',
-                                        help="Perform a build within a docker container. "
-                                             "NOTE: When the --docker_build flag is used, all arguments with paths "
-                                             "should be relative to the directory containing open-ce. Only files "
-                                             "within the open-ce directory and local_files will be visible at "
-                                             "build time."))
-
-    SKIP_BUILD_PACKAGES = (lambda parser: parser.add_argument(
-                                        '--skip_build_packages',
-                                        action='store_true',
-                                        help="Do not perform builds of packages."))
-
-    RUN_TESTS = (lambda parser: parser.add_argument(
-                                        '--run_tests',
-                                        action='store_true',
-                                        help="Run Open-CE tests for each potential conda environment"))
-
-    CONDA_ENV_FILE = (lambda parser: parser.add_argument(
-                                        '--conda_env_file',
-                                        type=str,
-                                        help='Location of conda environment file.' ))
-
-    LOCAL_CONDA_CHANNEL = (lambda parser: parser.add_argument(
-                                        '--local_conda_channel',
-                                        type=str,
-                                        default=DEFAULT_OUTPUT_FOLDER,
-                                        help='Path where built conda packages are present.'))
-
-    TEST_WORKING_DIRECTORY = (lambda parser: parser.add_argument(
-                                        '--test_working_dir',
-                                        type=str,
-                                        default="./",
-                                        help="Directory where tests will be executed."))
-
-
-def add_subparser(subparsers, command, arguments, *args, formatter_class=OpenCEFormatter, **kwargs):
-    '''
-    Make a parser from a list of OPEN-CE Arguments.
-    '''
-    subparser = subparsers.add_parser(command, *args, formatter_class=formatter_class, **kwargs)
-    for argument in arguments:
-        argument(subparser)
-    return subparser
-
-def parse_arg_list(arg_list):
-    ''' Turn a comma delimited string into a python list'''
-    if isinstance(arg_list, list):
-        return arg_list
-    return arg_list.split(",") if not arg_list is None else list()
-
 def make_variants(python_versions=DEFAULT_PYTHON_VERS, build_types=DEFAULT_BUILD_TYPES, mpi_types=DEFAULT_MPI_TYPES,
 cuda_versions=DEFAULT_CUDA_VERS):
     '''Create a cross product of possible variant combinations.'''
-    variants = { 'python' : parse_arg_list(python_versions),
-                 'build_type' : parse_arg_list(build_types),
-                 'mpi_type' :  parse_arg_list(mpi_types),
-                 'cudatoolkit' : parse_arg_list(cuda_versions)}
+    variants = { 'python' : inputs.parse_arg_list(python_versions),
+                 'build_type' : inputs.parse_arg_list(build_types),
+                 'mpi_type' :  inputs.parse_arg_list(mpi_types),
+                 'cudatoolkit' : inputs.parse_arg_list(cuda_versions)}
     return [dict(zip(variants,y)) for y in product(*variants.values())]
 
 def remove_version(package):
